Swap execution strategy in `index.js`.
When the function exported by `index.js` is called with itself, it
converts from command line execution to web service. When the web
service middleware constructor is invoked, it will latch the behavior
from command line execution over to web service. It cannot be toggled
back to command line execution.